dom/tests/mochitest/mouselock/test_movement.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=633602
-->
<head>
    <title>Test for Bug 633602</title>
    <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
    <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
    <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
    <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">Mozilla Bug 633602</a>
<div id="content">
</div>
    <pre id="test">
        <script type="application/javascript">

            /** Test for Bug 633602 **/
            SimpleTest.waitForExplicitFinish()

            SimpleTest.waitForFocus(function () {

                SpecialPowers.setBoolPref("full-screen-api.allow-trusted-requests-only", false);
                var pointer = navigator.pointer;
                var ePrevious = document.createEvent("MouseEvent");
                var eNow = document.createEvent("MouseEvent");
                var ep = document.createEvent("MouseEvent");
                var i = false;
                var t = 0;

                ePrevious.initMouseEvent("mousemove", true, false, window, 0, 0, 0, 0, 0, false, false, false, false, 0, null);
                eNow.initMouseEvent("mousemove", true, false, window, 0, 0, 0, 0, 0, false, false, false, false, 0, null);
                document.body.addEventListener("mousemove", function (e) {

                    if (t <= 3) {
                        //check if mouseEvent contains movementX/Y

                        ok(e.movementX != null, "MouseEvent must contain movementX.");
                        ok(e.movementY != null, "MouseEvent must contain movementY.");


                        eNow = e;
                        ePrevious = ep;
                        ep = e;

                        if (i) {    //the validation won't start at the first time

                            //check if movementX/Y = eNow.screenX/Y - ePrevious.screenX/Y
                            if (pointer.islocked() == true || pointer.islocked() == false) {  //movementX/Y must be valid regardless of mouse lock state
                                is(e.movementX, eNow.screenX - ePrevious.screenX, "movementX shoulde equal eNow.screenX - ePrevious.screenX");
                                is(e.movementY, eNow.screenY - ePrevious.screenY, "movementY shoulde equal eNow.screenY - ePrevious.screenY");
                            }


                        }
                        i = true;
                        SimpleTest.finish();
                        t++;
                    }
                }, false);
                document.body.addEventListener("mouseout", function (e) {
                    i = false;   //if the pointer goes out from the browser, reset
                }, false);
            }, false);
        </script>
    </pre>
    <p>move mouse over the page</p>
    <input type ="text" name ="txt" id="txt" />
</body>
</html>
